backend/requirements.txt
# AwakiFarmer MVP - Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# WhatsApp Integration
twilio==8.11.0

# AI/ML
anthropic==0.18.0
requests==2.31.0
pillow==10.2.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Configuration & Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing (optional for MVP)
pytest==7.4.4
httpx==0.26.0
